+++ src/infi/clickhouse_orm/migrations.py
@@ -6,6 +6,7 @@ from .engines import MergeTree
 from .utils import escape
 
 from six.moves import zip
+from six import iteritems
 
 import logging
 logger = logging.getLogger('migrations')
@@ -65,7 +66,7 @@ class AlterTable(Operation):
         table_fields = dict(self._get_table_fields(database))
 
         # Identify fields that were deleted from the model
-        deleted_fields = set(table_fields.keys()) - set(name for name, field in self.model_class._fields)
+        deleted_fields = set(table_fields.keys()) - set(self.model_class.fields())
         for name in deleted_fields:
             logger.info('        Drop column %s', name)
             self._alter_table(database, 'DROP COLUMN %s' % name)
@@ -73,7 +74,7 @@ class AlterTable(Operation):
 
         # Identify fields that were added to the model
         prev_name = None
-        for name, field in self.model_class._fields:
+        for name, field in iteritems(self.model_class.fields()):
             if name not in table_fields:
                 logger.info('        Add column %s', name)
                 assert prev_name, 'Cannot add a column to the beginning of the table'
@@ -89,7 +90,8 @@ class AlterTable(Operation):
         # The order of class attributes can be changed any time, so we can't count on it
         # Secondly, MATERIALIZED and ALIAS fields are always at the end of the DESC, so we can't expect them to save
         # attribute position. Watch https://github.com/Infinidat/infi.clickhouse_orm/issues/47
-        model_fields = {name: field.get_sql(with_default_expression=False) for name, field in self.model_class._fields}
+        model_fields = {name: field.get_sql(with_default_expression=False)
+                        for name, field in iteritems(self.model_class.fields())}
         for field_name, field_sql in self._get_table_fields(database):
             # All fields must have been created and dropped by this moment
             assert field_name in model_fields, 'Model fields and table columns in disagreement'