+++ tests/test_querysets.py
@@ -3,6 +3,7 @@ from __future__ import unicode_literals, print_function
 import unittest
 
 from infi.clickhouse_orm.database import Database
+from infi.clickhouse_orm.query import Q
 from .base_test_with_data import *
 import logging
 from datetime import date, datetime
@@ -58,6 +59,15 @@ class QuerySetTestCase(TestCaseWithData):
         self._test_qs(qs.filter(first_name__endswith='IA'), 0) # case sensitive
         self._test_qs(qs.filter(first_name__iendswith='ia'), 3) # case insensitive
         self._test_qs(qs.filter(first_name__iendswith=''), 100) # empty suffix
+
+    def test_filter_with_q_objects(self):
+        qs = Person.objects_in(self.database)
+        self._test_qs(qs.filter(Q(first_name='Ciaran')), 2)
+        self._test_qs(qs.filter(Q(first_name='Ciaran') | Q(first_name='Chelsea')), 3)
+        self._test_qs(qs.filter(Q(first_name__in=['Warren', 'Whilemina', 'Whitney']) & Q(height__gte=1.7)), 3)
+        self._test_qs(qs.filter((Q(first_name__in=['Warren', 'Whilemina', 'Whitney']) & Q(height__gte=1.7) |
+                                 (Q(first_name__in=['Victoria', 'Victor', 'Venus']) & Q(height__lt=1.7)))), 4)
+        self._test_qs(qs.filter(Q(first_name='Elton') & ~Q(last_name='Smith')), 1)
 
     def test_filter_unicode_string(self):
         self.database.insert([
@@ -324,6 +334,20 @@ class AggregateTestCase(TestCaseWithData):
         print(qs.as_sql())
         self.assertEquals(qs.count(), 1)
 
+    def test_double_underscore_field(self):
+        class Mdl(Model):
+            the__number = Int32Field()
+            the__next__number = Int32Field()
+            engine = Memory()
+        qs = Mdl.objects_in(self.database).filter(the__number=1)
+        self.assertEquals(qs.conditions_as_sql(), 'the__number = 1')
+        qs = Mdl.objects_in(self.database).filter(the__number__gt=1)
+        self.assertEquals(qs.conditions_as_sql(), 'the__number > 1')
+        qs = Mdl.objects_in(self.database).filter(the__next__number=1)
+        self.assertEquals(qs.conditions_as_sql(), 'the__next__number = 1')
+        qs = Mdl.objects_in(self.database).filter(the__next__number__gt=1)
+        self.assertEquals(qs.conditions_as_sql(), 'the__next__number > 1')
+
 
 Color = Enum('Color', u'red blue green yellow brown white black')
 
@@ -341,3 +365,5 @@ class SampleModel(Model):
 class Numbers(Model):
 
     number = UInt64Field()
+
+